Fixed corrupted git. Shooting and killing asteroids

+++ circleshape.py
@@ -3,7 +3,7 @@ from typing import Self
 
 # Base class for game objects
 class CircleShape(pygame.sprite.Sprite):
-    def __init__(self, x: int, y: int, radius: int):
+    def __init__(self, x: float, y: float, radius: int):
         # we will use this later
         if hasattr(self, "containers"):
             super().__init__(self.containers)
